+++ deliver.js
@@ -1,9 +1,7 @@
-/*jshint asi: true undef: true es5: true node: true devel: true
-         globalstrict: true forin: true latedef: false supernew: true */
-/*global define: true */
-
 "use strict";
 
-var Method = require('method')
+var method = require("method")
 // Method delivers pending value.
-module.exports = Method()
+var deliver = method("deliver")
+
+module.exports = deliver