+++ History.md
@@ -3,6 +3,7 @@
 ## 0.0.2 / 2012-11-10
 
   - Add browser integration tests.
+  - Update to new method library.
 
 ## 0.0.2 / 2012-10-15